Interface: add host, join(), repository, split(), sep
